% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SankeyShiny.R
\name{run_SankeyShiny}
\alias{run_SankeyShiny}
\title{Run the Sankey Shiny App}
\usage{
run_SankeyShiny()
}
\value{
A shiny app object
}
\description{
Launches the interactive Shiny application for Sankey combination plots.
}
